+++ libraries/base/changelog.md
@@ -6,6 +6,8 @@
 
   * Add reverse application operator `Data.Function.(&)`
 
+  * Add `Data.List.sortOn` sorting function
+
 ## 4.7.0.0  *Apr 2014*
 
   * Bundled with GHC 7.8.1